# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UniStore API",
        "description": "API REST UniStore - BIT",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/panier": {
            "get": {
                "tags": [
                    "Panier"
                ],
                "summary": "Contenu du panier",
                "operationId": "29eb6c162caa1a7322a6c0e555d692f9",
                "responses": {
                    "200": {
                        "description": "Panier récupéré"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Panier"
                ],
                "summary": "Ajouter au panier",
                "operationId": "76f05419379182829ca9c21f6171db18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "produit_id",
                                    "quantite"
                                ],
                                "properties": {
                                    "produit_id": {
                                        "type": "integer"
                                    },
                                    "quantite": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ajout réussi"
                    },
                    "422": {
                        "description": "Erreur de validation ou stock"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Panier"
                ],
                "summary": "Vider le panier",
                "operationId": "86dade00d70bcc4fc3ccbb3df04813d6",
                "responses": {
                    "200": {
                        "description": "Panier vidé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/panier/{id}": {
            "put": {
                "tags": [
                    "Panier"
                ],
                "summary": "Modifier quantité",
                "operationId": "5b1b57d9d2b6f38eaa3872aa22e6a624",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantite": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour réussie"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Panier"
                ],
                "summary": "Supprimer un article",
                "operationId": "287a79f560a1c2b7d520d6ddf4acdd9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article supprimé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}